requirements.txt
mindsdb_sdk
python-dotenv
Flask
requests
reportlab